fix: inject correct server base URL
We are currently using `get_current_site().domain` as the server base URL
displayed in the mail to the user who requested an OTP. This doesn't
properly display the correct port – which is problematic if different
from standard port. Also, the mail templates are currently forcing
`https` protocol and shouldn't.

This present commit fixes these two problems.